[IMP] ni_device,ni_observation: update device usage
