+++ ni_device/__manifest__.py
@@ -14,7 +14,7 @@
         "security/ir.model.access.csv",
         "security/ir_rule_data.xml",
         "data/ir_sequence_data.xml",
-        "views/ni_device_usage_logs_views.xml",
+        "views/ni_device_usage_views.xml",
         "views/ni_device_views.xml",
         "views/ni_device_type_views.xml",
         "views/ni_device_request_views.xml",
@@ -24,6 +24,7 @@
         "views/ni_device_label_layout_views.xml",
         "views/ni_observation_views.xml",
         "views/ni_observation_sheet_views.xml",
+        "views/ni_patient_smartcard_views.xml",
         "views/ni_device_menu.xml",
         "report/device_label_layout_template.xml",
     ],
