+++ ni_device/data/ir_sequence_data.xml
@@ -33,5 +33,12 @@
             <field name="padding">4</field>
             <field name="company_id" eval="False" />
         </record>
+        <record id="seq_ni_device_usage" model="ir.sequence">
+            <field name="name">Device Usage</field>
+            <field name="code">ni.device.usage</field>
+            <field name="prefix">DU-%(year)s%(month)s</field>
+            <field name="padding">6</field>
+            <field name="company_id" eval="False" />
+        </record>
     </data>
 </odoo>
