+++ ni_device/models/__init__.py
@@ -6,7 +6,6 @@ from . import ni_device_request
 from . import ni_device_holder
 from . import ni_device_repair
 from . import ni_device_usage
-from . import ni_device_usage_log
 from . import ni_device_usage_log_mixin
 from . import ni_device_dispose_type
 from . import ni_device_label_layout
