+++ ni_device/models/ni_observation.py
@@ -5,7 +5,4 @@ from odoo import fields, models
 class Observation(models.Model):
     _inherit = "ni.observation"
 
-    device_id = fields.Many2one("ni.device", index=True, ondelete="restrict")
-    usage_log_id = fields.Many2one(
-        "ni.device.usage.log", index=True, ondelete="cascade"
-    )
+    device_id = fields.Many2one("ni.device", index=True)
