+++ ni_device/models/ni_observation_sheet.py
@@ -5,4 +5,4 @@ class ObservationSheet(models.Model):
     _name = "ni.observation.sheet"
     _inherit = ["ni.observation.sheet", "ni.device.usage.log.mixin"]
 
-    device_id = fields.Many2one("ni.device", index=True, ondelete="restrict")
+    device_id = fields.Many2one("ni.device", index=True)
