+++ ni_device/models/ni_patient_smartcard.py
@@ -1,5 +1,8 @@
-from odoo import models
+from odoo import fields, models
 
 
 class PatientSmartcard(models.Model):
-    _inherit = ["ni.patient.smartcard"]
+    _name = "ni.patient.smartcard"
+    _inherit = ["ni.patient.smartcard", "ni.device.usage.log.mixin"]
+
+    device_id = fields.Many2one("ni.device", index=True)
