+++ ni_device/security/ir.model.access.csv
@@ -14,5 +14,3 @@ access_ni_device_dispose_type_admin,access_ni_device_dispose_type_admin,model_ni
 access_ni_device_label_layout,access_ni_device_label_layout,model_ni_device_label_layout,ni_patient.group_user,1,1,1,1
 access_ni_device_usage,access_ni_device_usage,model_ni_device_usage,ni_patient.group_viewer,1,0,0,0
 access_ni_device_usage_user,access_ni_device_usage_user,model_ni_device_usage,ni_patient.group_user,1,1,1,1
-access_ni_device_usage_log,access_ni_device_usage_log,model_ni_device_usage_log,ni_patient.group_viewer,1,0,0,0
-access_ni_device_usage_log_user,access_ni_device_usage_log_user,model_ni_device_usage_log,ni_patient.group_user,1,1,1,1
