+++ ni_device/views/ni_device_menu.xml
@@ -107,12 +107,12 @@
         action="ni_device.ni_device_request_action"
     />
     <menuitem
-        id="menu_device_admin_usage_log_all"
+        id="menu_device_admin_usage_all"
         name="ประวัติการใช้งานอุปกรณ์ทั้งหมด"
         groups="ni_patient.group_admin"
         parent="menu_device_admin_root"
         sequence="2"
-        action="ni_device.ni_device_usage_log_action"
+        action="ni_device.ni_device_usage_action"
     />
     <menuitem
         id="configuration_menu"
