+++ ni_device/views/ni_device_views.xml
@@ -164,13 +164,13 @@
                         groups="ni_patient.group_user"
                     >
                         <button
-                            name="ni_device.ni_device_usage_log_action_this"
-                            attrs="{'invisible': [('usage_log_count', '=', 0)]}"
+                            name="ni_device.ni_device_usage_action_this"
+                            attrs="{'invisible': [('usage_count', '=', 0)]}"
                             class="oe_stat_button"
                             icon="fa-list-alt"
                             type="action"
                         >
-                            <field name="usage_log_count" string="ประวัติใช้งาน" widget="statinfo" />
+                            <field name="usage_count" string="ประวัติใช้งาน" widget="statinfo" />
                         </button>
                         <button
                             name="ni_device.ni_device_repair_action_this"
