+++ ni_device/views/ni_observation_sheet_views.xml
@@ -13,6 +13,16 @@
             <xpath expr="//field[@name='occurrence']" position="after">
                 <field name="device_id" />
             </xpath>
+            <!-- เพิ่มปุ่มใน header -->
+            <xpath expr="//header" position="inside">
+                <button
+                    name="action_create_device_log"
+                    type="object"
+                    string="Create Device Usage"
+                    icon="fa-bug"
+                    groups="base.group_no_one"
+                />
+            </xpath>
         </field>
     </record>
 </odoo>
