+++ ni_observation/views/ni_observation_sheet_views.xml
@@ -82,6 +82,7 @@
             <form>
                 <field name="id" invisible="True" />
                 <field name="active" invisible="True" />
+                <header />
                 <sheet>
                     <div name="button_box" class="oe_button_box" groups="base.group_user" />
                     <widget
